refactor(components): extract geo-utils to break MapView cycle
sensor-utils.ts ↔ array-offset.ts had a real runtime cycle: sensor-utils
called computeArrayCentre (from array-offset), and array-offset called
three geo primitives (geodesicDestination, interpolateTrackPosition,
interpolateTrackCourse) back from sensor-utils.

Extracted the three pure primitives into shared/components/src/MapView/
geo-utils.ts. array-offset.ts now imports from geo-utils directly.
sensor-utils.ts also imports from geo-utils and re-exports the three
names, so tests and any external consumers that import these helpers
from sensor-utils keep working (zero public-API change).

Removed the now-unused local EARTH_RADIUS constant and findNearestPointIndex
import from sensor-utils.

madge --circular shared/components/src: 1 cycle → 0.
Checks: pnpm build, pnpm -r typecheck, pnpm lint (0 errors, warnings
unchanged), 1468 vitest tests pass.
