+++ shared/components/src/MapView/array-offset.ts
@@ -14,7 +14,7 @@
  */
 
 import type { SensorData } from '@debrief/schemas';
-import { geodesicDestination, interpolateTrackCourse, interpolateTrackPosition } from './sensor-utils';
+import { geodesicDestination, interpolateTrackCourse, interpolateTrackPosition } from './geo-utils';
 import { findNearestPointIndex } from './temporal-utils';
 
 /** Mean Earth radius in metres (matches Python haversine_distance_metres) */
