+++ shared/components/src/MapView/sensor-utils.ts
@@ -14,15 +14,20 @@
 import type { SensorContact, SensorData, TrackFeature } from '@debrief/schemas';
 import type { DisplayMode } from '../utils/types';
 import { computeArrayCentre } from './array-offset';
-import { findNearestPointIndex } from './temporal-utils';
+import {
+  geodesicDestination,
+  interpolateTrackCourse,
+  interpolateTrackPosition,
+} from './geo-utils';
+
+// Re-export geo primitives for backward compatibility with existing consumers
+// (tests and call sites that import these helpers from sensor-utils).
+export { geodesicDestination, interpolateTrackCourse, interpolateTrackPosition };
 
 // ── Constants ───────────────────────────────────────────────────────
 
 /** Maximum bearing line extent when no range is specified (5 degrees of latitude in metres) */
 export const MAXIMUM_SENSOR_BEARING_RANGE = 5 * 111_120; // ~555,600 metres
-
-/** Earth radius in metres for haversine calculations */
-const EARTH_RADIUS = 6_371_000;
 
 /** Factor for darkening colours (matches Java Color.darker()) */
 const DARKEN_FACTOR = 0.7;
@@ -157,45 +162,6 @@ export function calculateSnailProportion(
 // ── Bearing Geometry ────────────────────────────────────────────────
 
 /**
- * Calculate the destination point given start point, bearing, and distance.
- * Uses haversine formula for geodesic accuracy.
- *
- * @param origin [lon, lat] starting point
- * @param bearing Degrees from north (0-360)
- * @param distanceMetres Distance in metres
- * @returns [lon, lat] destination point
- */
-export function geodesicDestination(
-  origin: [number, number],
-  bearing: number,
-  distanceMetres: number,
-): [number, number] {
-  const [lon, lat] = origin;
-  const toRad = Math.PI / 180;
-  const toDeg = 180 / Math.PI;
-
-  const lat1 = lat * toRad;
-  const lon1 = lon * toRad;
-  const brng = bearing * toRad;
-  const d = distanceMetres / EARTH_RADIUS;
-
-  const sinD = Math.sin(d);
-  const cosD = Math.cos(d);
-  const sinLat1 = Math.sin(lat1);
-  const cosLat1 = Math.cos(lat1);
-
-  const lat2 = Math.asin(sinLat1 * cosD + cosLat1 * sinD * Math.cos(brng));
-  const lon2 =
-    lon1 +
-    Math.atan2(
-      Math.sin(brng) * sinD * cosLat1,
-      cosD - sinLat1 * Math.sin(lat2),
-    );
-
-  return [lon2 * toDeg, lat2 * toDeg];
-}
-
-/**
  * Calculate the far end of a bearing line.
  * If range is provided, uses it directly.
  * If no range, extends to MAXIMUM_SENSOR_BEARING_RANGE (5 degrees latitude).
@@ -207,71 +173,6 @@ export function computeBearingFarEnd(
 ): [number, number] {
   const distance = range ?? MAXIMUM_SENSOR_BEARING_RANGE;
   return geodesicDestination(origin, bearing, distance);
-}
-
-// ── Track Position Interpolation ────────────────────────────────────
-
-/**
- * Interpolate the host track's position at a given timestamp.
- * Uses binary search + linear interpolation on the positions/coordinates arrays.
- *
- * @param coordinates Array of [lon, lat] from track geometry
- * @param positions Array of { time: string } from track properties
- * @param targetTimeMs Target timestamp (epoch ms)
- * @returns [lon, lat] interpolated position, or null if time is out of range
- */
-export function interpolateTrackPosition(
-  coordinates: [number, number][],
-  positions: Array<{ time: string }>,
-  targetTimeMs: number,
-): [number, number] | null {
-  if (coordinates.length === 0 || positions.length === 0) return null;
-  if (coordinates.length !== positions.length) return null;
-
-  const timestamps = positions.map((p) => Date.parse(p.time));
-
-  // Out of range
-  if (targetTimeMs < timestamps[0]! || targetTimeMs > timestamps[timestamps.length - 1]!) {
-    return null;
-  }
-
-  // Find nearest index using binary search
-  const idx = findNearestPointIndex(timestamps, targetTimeMs);
-
-  // Exact match
-  if (timestamps[idx] === targetTimeMs) {
-    return coordinates[idx]!;
-  }
-
-  // Determine which pair to interpolate between
-  let lowIdx: number;
-  let highIdx: number;
-  if (timestamps[idx]! < targetTimeMs) {
-    lowIdx = idx;
-    highIdx = idx + 1;
-  } else {
-    lowIdx = idx - 1;
-    highIdx = idx;
-  }
-
-  // Bounds check
-  if (lowIdx < 0 || highIdx >= timestamps.length) {
-    return coordinates[idx]!;
-  }
-
-  const t0 = timestamps[lowIdx]!;
-  const t1 = timestamps[highIdx]!;
-  const dt = t1 - t0;
-  if (dt === 0) return coordinates[lowIdx]!;
-
-  const fraction = (targetTimeMs - t0) / dt;
-  const [lon0, lat0] = coordinates[lowIdx]!;
-  const [lon1, lat1] = coordinates[highIdx]!;
-
-  return [
-    lon0 + (lon1 - lon0) * fraction,
-    lat0 + (lat1 - lat0) * fraction,
-  ];
 }
 
 // ── Colour Inheritance ──────────────────────────────────────────────
@@ -301,28 +202,6 @@ export function getRelativeBearing(courseDeg: number, bearingDeg: number): numbe
   while (rel > 180) rel -= 360;
   while (rel < -180) rel += 360;
   return rel;
-}
-
-/**
- * Interpolate the host track's course at a given timestamp.
- * Returns course in degrees, or null if time is out of range.
- */
-export function interpolateTrackCourse(
-  positions: Array<{ time: string; course?: number }>,
-  targetTimeMs: number,
-): number | null {
-  if (positions.length === 0) return null;
-
-  const timestamps = positions.map((p) => Date.parse(p.time));
-
-  // Out of range
-  if (targetTimeMs < timestamps[0]! || targetTimeMs > timestamps[timestamps.length - 1]!) {
-    return null;
-  }
-
-  const idx = findNearestPointIndex(timestamps, targetTimeMs);
-  const pos = positions[idx];
-  return pos?.course ?? null;
 }
 
 /**
